Big refactor to work with backdrop-collector package
This incorporates the backdrop-collector package and moves everything
for this collector into one module.

The top level package name as been changed so that is doesn't conflict
with that of backdrop-collector.

The realtime module now encapsulates everything we need to know about
collecting from that source. It exposes a Collector class with a well
defined interface. Also, the environment specific and private details of the target
(the full url and the token) have been extracted out into a separate
file.

The idea is that we don't have to have multiple collector projects that
are deployed separately and cause us much maintenance
pain. We have one and each collector is in it's own module. The
`collect.py` shows how we could integrate the current with the
`backdrop-collector` package and `collect_any.py` shows how we might
extend this to load the right module depending on a field in the query
file. Different collector modules could still be (and probably should
be) maintained in their own repositories and installed as dependencies.

What are peoples thoughts on this?