fix error, clean up tests - timezone issue should now be fixed
As for as we can tell, the process for generating realtime timestamps
used to go

generate timestamp now for utc e.g. 10:20 am +00:00-> change timezone to
gmt without affecting the rest of the time e.g. 10:20 am +01:00 -> send
to backdrop -> convert back to utc without affecting the rest of the
time e.g. 10:20 am +00:00 -> save, timestamps as they are now will be
sent to spotlight.

This puts the time coming to spotlight one hour back vs local time

The fix:

generate timestamp now for gmt e.g. 11:20 am +01:00 when BST -> send to
backdrop -> convert back to utc without affecting the rest of the time
e.g. 11:20 am +00:00 -> save, timestamps as they are now will be sent to
spotlight.

Tests force the system time to a summer time and winter one and check we
are sending the correct timestamp to backdrop...